Changed AHRS example to LSM because MPU on a RPI failed

+++ Navio/Extra/C++_Examples/AHRS/AHRS.hpp
@@ -30,7 +30,7 @@ public:
     void updateIMU(float dt);
     void setGyroOffset();
     void getEuler(float* roll, float* pitch, float* yaw);
-    void read_gyroscope(float* gx,float* gy,float* gz);
+    void read_gyroscope(double* gx,double* gy,double* gz);
 
     float invSqrt(float x);
     float getW();
